requirement.txt
amqp==2.6.1
arabic-reshaper==2.1.1
asgiref==3.3.1
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
billiard==3.6.3.0
celery==4.4.7
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coverage==4.5.1
cryptography==2.3.1
diff-match-patch==20121119
dj-database-url==0.5.0
django-ckeditor
Django==2.0.2
django-appconf==1.0.2
django-celery-results==2.0.0
django-chosen==0.1
django-embed-video
django-extra-views==0.11.0
django-haystack==2.8.1
django-import-export==1.0.1
django-js-asset==1.2.2
django-mathfilters==0.4.0
django-multiselectfield==0.1.12
django-nose==1.4.5
django-oscar==1.6.2
django-oscar-payu==1.1
django-payu==0.5
django-phonenumber-field==2.0.0
django-ranged-response==0.2.0
django-select2==6.1.0
django-simple-captcha==0.5.9
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.2
django-wkhtmltopdf==3.3.0
docopt==0.6.2
envdir==1.0.1
et-xmlfile==1.0.1
fabric==2.3.1
factory-boy==2.11.1
Faker==0.8.15
fhurl==0.1.10
future==0.18.2
html5lib==1.1
idna==2.7
importd==0.4.3
importlib-metadata==2.1.1
invoke==1.1.1
jdcal==1.4
kombu==4.6.11
mock==2.0.0
nose==1.3.7
num2words==0.5.10
odfpy==1.3.6
openpyxl==2.5.4
paramiko==2.4.1
path.py==11.0.1
pbr==4.0.4
pdfkit==0.6.1
phonenumbers==8.9.7
phonenumberslite==8.9.7
Pillow==5.2.0
psycopg2-binary==2.7.4
purl==1.4
pyasn1==0.4.4
pycountry==18.5.26
pycparser==2.18
pycrypto==2.6.1
pycryptodomex==3.9.4
PyNaCl==1.2.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
redis==3.5.3
reportlab==3.5.50
requests==2.19.1
six
smarturls==0.1.9
sorl-thumbnail==12.4.1
speaklater==1.3
sqlparse==0.4.1
tablib==0.12.1
text-unidecode==1.2
unicodecsv==0.14.1
Unidecode==1.0.22
urllib3==1.23
vine==1.3.0
webencodings==0.5.1
xlrd==1.1.0
xlwt==1.3.0
yolk==0.4.3
zipp==1.2.0
